Separate complete and partial insertions
Previoulsy, we recorded all language box insertions as one statistic.
However, there is a big difference between a complete and a partial
insertion of a language box. With this commit we log the length of the
inserted language box as well as the length of the inserted fragment, so
we can distinguish between complete and partial insertions in the
tables.